Wrong range Assumption
Test IP address from Telecom Argentina S.A. expected wrong data.
LACNIC answers in shortened CIDR form 190.190/15 which in fact is 190.190.0.0/15 and not the wronly assumed 190.0.0.190/15 (which does not include the input parameter IP).
Wrong assumption due to seemingly wrong implementation by the framework to analyze 190.190/15 correctly. The expected false results were only met because in fact the method by which the AddressRange is analyzed returns false data.